backend/requirements.txt
django
channels
djangorestframework
channels_redis  # Optional for production